requirements-core.txt
Package                       Version
----------------------------- -------
dj-rest-auth                  7.0.1
django-cors-headers           4.7.0
django-filter                 25.1
django-guardian               3.0.0
djangorestframework_simplejwt 5.5.0
gunicorn                      23.0.0
Markdown                      3.8
pip                           25.1.1
psycopg2                      2.9.10
Pygments                      2.19.1
python-dotenv                 1.1.0
setuptools                    65.5.0
